LayoutTests/http/tests/permissions/permissions-query-window-expected.txt

This test checks that Permissions::query() works in window contexts

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS state is "prompt"
PASS state is "granted"
PASS state is "prompt"
PASS successfullyParsed is true

TEST COMPLETE
